+++ java/javasrc/src/main/java/com/longbridge/SdkNative.java
@@ -451,6 +451,14 @@ public class SdkNative {
                         Object opts,
                         AsyncCallback callback);
 
+        public static native void fundamentalContextMacroeconomicIndicators(long context,
+                        Object country, Object keyword, Object offset, Object limit,
+                        AsyncCallback callback);
+
+        public static native void fundamentalContextMacroeconomic(long context,
+                        Object indicatorCode, Object startTime, Object endTime, Object offset, Object limit,
+                        AsyncCallback callback);
+
         public static native void portfolioContextProfitAnalysisFlows(long context, Object opts,
                         AsyncCallback callback);