+++ nodejs/src/fundamental/context.rs
@@ -287,4 +287,39 @@ impl FundamentalContext {
             .map_err(ErrorNewType)?
             .into())
     }
+
+    /// List macroeconomic indicators
+    #[napi]
+    pub async fn macroeconomic_indicators(
+        &self,
+        country: Option<MacroeconomicCountry>,
+        keyword: Option<String>,
+        offset: Option<i32>,
+        limit: Option<i32>,
+    ) -> Result<MacroeconomicIndicatorListResponse> {
+        Ok(self
+            .ctx
+            .macroeconomic_indicators(country.map(Into::into), keyword, offset, limit)
+            .await
+            .map_err(ErrorNewType)?
+            .into())
+    }
+
+    /// Get historical data for a macroeconomic indicator
+    #[napi]
+    pub async fn macroeconomic(
+        &self,
+        indicator_code: String,
+        start_date: Option<String>,
+        end_date: Option<String>,
+        offset: Option<i32>,
+        limit: Option<i32>,
+    ) -> Result<MacroeconomicResponse> {
+        Ok(self
+            .ctx
+            .macroeconomic(indicator_code, start_date, end_date, offset, limit)
+            .await
+            .map_err(ErrorNewType)?
+            .into())
+    }
 }